backend/requirements.txt
flask
flask-cors
tensorflow
pillow
pymongo
python-dotenv
google-generativeai